Moving Pathway of the Damned to Github.
Added readme.markdown with history of Pathway, and some installation instructions.

Updated Pathway of the Damned to the latest ArcEmu revision as of today.

No data was changed except for the two ArcEmu updates in between the time of the last commit on Assembla and the move to Github.

See readme.markdown for my next moves in development with this database.